laddertools: ranking algorithm refactoring
