+++ requirements.txt
@@ -5,3 +5,4 @@ numpy==1.23.4
 pre-commit==2.20.0
 pydantic==1.10.2
 PyYAML==6.0
+trueskill==0.4.5
